fix cfilter.Insert(...) rebalancing
Fixes #8.
The modulo operator was incorrectly being applied to the result of
the hashfunction, the result of which was then XOR-ed to be the new search
index. This opened up the possibility for the index to be greater than the
filter size resulting in an index out of range panic.